move readme endpoint types to api folder
