+++ frontend/src/api/api.ts
@@ -3,21 +3,19 @@ import {
   convertAnalysisResults,
   convertConvertIds,
   revertAnalysisInputs,
-  type _AnalysisResults,
-  type _ConvertIds,
   type AnalysisInputs,
-  type Species,
+} from "@/api/convert";
+import {
+  type _AnalysisResults,
+  type _ConvertIdsInputs,
+  type _ConvertIdsResults,
 } from "@/api/types";
 
 /** check input list of genes. convert to entrez, check if in-network, etc. */
-export const checkGenes = async (
-  genes: string[],
-  species: Species = "Human",
-) => {
+export const checkGenes = async (params: _ConvertIdsInputs) => {
   const headers = new Headers();
   headers.append("Content-Type", "application/json");
-  const params = { genes, species };
-  const response = await request<_ConvertIds>(
+  const response = await request<_ConvertIdsResults>(
     `${api}/gpz-convert-ids`,
     undefined,
     { method: "POST", headers, body: JSON.stringify(params) },
