+++ frontend/src/pages/analysis/Network.tsx
@@ -17,7 +17,7 @@ import clsx from "clsx";
 import * as d3 from "d3";
 import { clamp, cloneDeep, truncate } from "lodash";
 import { useElementSize } from "@reactuses/core";
-import type { AnalysisInputs, AnalysisResults } from "@/api/types";
+import type { AnalysisInputs, AnalysisResults } from "@/api/convert";
 import Button from "@/components/Button";
 import CheckBox from "@/components/CheckBox";
 import Flex from "@/components/Flex";
