+++ frontend/src/pages/analysis/Predictions.tsx
@@ -1,4 +1,4 @@
-import type { AnalysisResults } from "@/api/types";
+import type { AnalysisResults } from "@/api/convert";
 import Exponential from "@/components/Exponential";
 import Table from "@/components/Table";
